JenkinScripts/Azure_Fetch_Instances/requirements.txt
azure-identity==1.15.0
azure-mgmt-compute==30.3.0
azure-mgmt-network==25.1.0
paramiko==3.3.1
scp==0.14.5
